+++ api/src/app.ts
@@ -44,6 +44,7 @@ app.use(deserializeTokens);
 
 // Application-level
 app.use("/", (req, res, next) => {
+  // console.clear();
   console.group("Application-level middleware running...");
   console.log("req.originalUrl:", req.originalUrl);
   console.groupEnd();