+++ api/src/middleware/deserializeTokens.ts
@@ -2,8 +2,6 @@ import { RequestHandler } from "express";
 
 // Extract tokens from request headers and cookies
 const deserializeTokens: RequestHandler = async (req, res, next) => {
-  console.clear();
-  console.log("req.headers['authorization']:", req.headers["authorization"]);
   const bearerHeader = req.headers["authorization"];
   const accessToken = bearerHeader && bearerHeader.split(" ")[1];
   const { refreshToken } = req.cookies;