+++ api/src/routes/routes.ts
@@ -1,14 +1,12 @@
-import express, { Application } from "express";
+import { Application } from "express";
 import userRoutes from "./user.route";
 import pizzaRoutes from "./pizza.route";
-import postRoutes from "./post.route";
 import authRoutes from "./auth.route";
 
 // Import routes and mount routes on specific paths
 const routes = (app: Application) => {
   app.use("/pizza", pizzaRoutes());
-  app.use("/user", userRoutes());
-  app.use("/posts", postRoutes());
+  app.use("/users", userRoutes());
   app.use("/auth", authRoutes());
 };