+++ api/src/utils/jwt.utils.ts
@@ -6,10 +6,6 @@ export const signJWT = async (
 ): Promise<string | null> => {
   return await new Promise((res, rej) => {
     jwt.sign(payload, "secretKey", options, (err, token) => {
-      console.log("signJWT");
-      console.log("token in signJWT:", token);
-      console.log("!token:", !token);
-      console.log("typeof token:", typeof token);
       if (err || !token) return res(null);
       return res(token);
     });
@@ -27,10 +23,6 @@ export const verifyJWT = async (
   token: string | null | undefined
 ): Promise<{ payload: null | JwtPayload; expired: boolean }> => {
   return await new Promise((res, rej) => {
-    console.log("verifyJWT");
-    console.log("token in verifyJWT:", token);
-    console.log("!token:", !token);
-    console.log("typeof token:", typeof token);
     if (!token) {
       return res({
         payload: null,