+++ cms/src/features/posts/components/PostCard.tsx
@@ -8,16 +8,20 @@ interface Props {
 
 const PostCard: React.FC<Props> = ({ data }) => {
   const { id, title, titleSlug, createdAt } = data;
-  const { author } = useParams();
-  // const url = encodeURI(title);
+  const { username } = useParams();
+  const postPath = `/${username}/${titleSlug}`;
   // TODO
   // Add
   //  Delete button
   //  Anchor heading
   return (
     <li>
-      <NavAnchor pathname={`/${author}/${titleSlug}`} textContent={title} />
+      <NavAnchor pathname={postPath} textContent={title} />
       <p>Posted on {createdAt.toString()}</p>
+      <NavAnchor
+        pathname={`/${username}/edit?goto=${postPath}`}
+        textContent="edit"
+      />
     </li>
   );
 };