+++ cms/src/interfaces/user.ts
@@ -10,4 +10,12 @@ export interface User {
   // comments?: Comment[] | null;
 }
 
+export interface AuthenticatedUser {
+  id: string;
+  role: Role;
+  username: string;
+  timestamp: Date;
+  blog: Blog | null;
+}
+
 type Role = "USER" | "AUTHOR" | "ADMIN";